cast the step to int

+++ pyglow.py
@@ -67,12 +67,12 @@ class PyGlow:
     def pulse_loop(self, led, value):
         step = 0
         while (step <= value):
-            self.set_leds(led, ceil(sin(step/value)))
+            self.set_leds(led, int(ceil(sin(step/value))))
             self.update_leds()
             sleep(0.005)
             step += 1
         while (step >= 0):
-            self.set_leds(led, ceil(sin(step/value)))
+            self.set_leds(led, int(ceil(sin(step/value))))
             self.update_leds()
             sleep(0.005)
             step -= 1
